demo

+++ lavis/models/blip2_models/blip2_t5.py
@@ -105,7 +105,7 @@ class Blip2T5(Blip2Base):
         inputs_t5 = self.t5_proj(query_output.last_hidden_state)
         atts_t5 = torch.ones(inputs_t5.size()[:-1], dtype=torch.long).to(image.device)
 
-        with torch.cuda.amp.autocast(dtype=torch.bfloat16):
+        with torch.cuda.amp.autocast(dtype=torch.float16):
 
             input_tokens = self.t5_tokenizer(
                 samples["text_input"],
@@ -205,7 +205,7 @@ class Blip2T5(Blip2Base):
             
         device_type = "cuda" if "cuda" in str(self.device) else "cpu"
         with torch.amp.autocast(
-            device_type=device_type, dtype=torch.bfloat16
+            device_type=device_type, dtype=torch.float16
         ):
             inputs_embeds = self.t5_model.encoder.embed_tokens(input_tokens.input_ids)
             inputs_embeds = torch.cat([inputs_t5, inputs_embeds], dim=1)
